refactor(shared): decompose ObservationManager into ObservationRegistry + ObservationEmitter
